refactor(observability): extract metrics + the Sentry scrubber — D1 part 3 of 3 (KAN-415)
src/lib/{metrics,sentry-scrub}.ts -> src/modules/observability/

TWO FILES, NOT FOUR. The observability module also declares src/app/api/health/
and src/app/status/, and those stay exactly where they are: in the App Router a
route's URL IS its directory path, so moving them would change /api/health and
/status — the two endpoints every uptime check, smoke test and promote gate
gets its answer from. Same reason the oauth-as extraction left its routes in
place. The module now spans two directories, which is the honest shape of a
Next.js module rather than a defect to tidy away.

THE INTERESTING PART IS WHERE THE IMPORTERS WERE. A src/ + tests/ sweep — the
one used for platform and guards — finds TWO of the four:

  src/app/admin/monitoring/page.tsx        @/lib/metrics
  tests/unit/{metrics-admin-window,sentry-scrub}.test.ts

The other two live at the REPO ROOT, outside src/ and tests/ entirely:

  instrumentation-client.ts:28   a normal static import
  instrumentation.ts:36-37       await import(
                                   '@/lib/sentry-scrub'
                                 )

That second one is a MULTI-LINE dynamic import whose specifier sits alone on
line 37, with no 'import' keyword anywhere near it. A line-oriented rewrite
sees nothing to match. This rewrite scans the whole tracked tree and operates
on file contents rather than lines, so both were caught — and tsc is the
backstop underneath, because TypeScript resolves a dynamic import specifier
when it is a string literal.

Those two files are the entire Sentry wiring. Miss them and scrubSentryEvent
stops being installed as beforeSend — the SEC-55 control that strips OAuth
secrets and PII from events before they leave the process. It would not have
failed a test: sentry-scrub.test.ts exercises the functions directly, so the
scrubber would keep passing its own suite while no longer being attached to
anything. Same shape as BUGS-85.

ESTATE: modules.json, kan416-derive.py, the test-path manifest.

docs/modularisation/KAN-414-F6-threading-fallout.md is NOT rewritten and joins
ARCHIVE_FILES. Its own second paragraph says 'the threading was then reverted,
so this document — not the code — is the deliverable', and it is a measurement
dated 2026-07-29. Rewriting paths inside a measurement restates the measurement.

No founder-owned path is touched: the only .tsx in the diff is
src/app/admin/monitoring/page.tsx, and src/app/admin/** is a documented
carve-out in check-ui-copy-ownership.sh.

261 suites / 3171 tests, tsc clean, all eight guards exit 0 — including CTL-041,
added in part 2, which is what confirms modules.json still describes the tree.

This completes sequencing D1. platform + guards + observability now exist as
named modules, which is what makes every later boundary rule expressible.
